4. Keeping Knowledge Organized with Indexes/requirements.txt
aioboto3==12.4.0
aiobotocore==2.12.3
aiohttp==3.9.5
aioitertools==0.11.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
boto3==1.34.69
botocore==1.34.69
cachetools==5.3.3
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cssselect==1.2.0
dataclasses-json==0.6.5
decorator==5.1.1
deepdiff==7.0.1
deeplake==3.9.3
dill==0.3.8
distro==1.9.0
emoji==2.11.1
executing==2.0.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.14.0
filetype==1.2.0
frozenlist==1.4.1
google-api-core==2.19.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.63.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
humbug==0.3.2
idna==3.7
install==1.3.5
ipython==8.24.0
jedi==0.19.1
jieba3k==0.35.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpickle==3.0.4
jsonpointer==2.4
langchain==0.1.17
langchain-community==0.0.36
langchain-core==0.1.50
langchain-google-community==1.0.3
langchain-openai==0.1.6
langchain-text-splitters==0.0.1
langdetect==1.0.9
langsmith==0.1.54
lxml==5.2.1
lxml_html_clean==0.1.1
lz4==4.3.3
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib-inline==0.1.7
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.25.2
ordered-set==4.1.0
orjson==3.10.3
outcome==1.3.0.post0
packaging==23.2
parso==0.8.4
pathos==0.3.2
pexpect==4.9.0
pillow==10.2.0
pox==0.3.4
ppft==1.7.6.8
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
pypdf==4.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
pyvis==0.3.2
PyYAML==6.0.1
rapidfuzz==3.9.0
regex==2024.4.28
requests==2.31.0
requests-file==2.0.0
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.10.1
selenium==4.20.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.30
stack-data==0.6.3
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.6.0
tinysegmenter==0.3
tldextract==5.1.2
tqdm==4.66.4
traitlets==5.14.3
trio==0.25.0
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.11.0
unstructured==0.11.8
unstructured-client==0.22.0
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
